remove camera options for franka
