--- source/isaaclab_tasks/isaaclab_tasks/core/dexsuite/config/franka/camera_cfg.py
@@ -1,244 +1,0 @@
-# Copyright (c) 2022-2026, The Isaac Lab Project Developers (https://github.com/isaac-sim/IsaacLab/blob/main/CONTRIBUTORS.md).
-# All rights reserved.
-#
-# SPDX-License-Identifier: BSD-3-Clause
-
-from dataclasses import MISSING
-
-import isaaclab.sim as sim_utils
-from isaaclab.managers import ObservationGroupCfg as ObsGroup
-from isaaclab.managers import ObservationTermCfg as ObsTerm
-from isaaclab.managers import SceneEntityCfg
-from isaaclab.sensors import CameraCfg, MultiMeshRayCasterCameraCfg, patterns
-from isaaclab.utils.configclass import configclass
-from isaaclab.utils.noise import UniformNoiseCfg as Unoise
-
-from isaaclab_tasks.utils import PresetCfg
-from isaaclab_tasks.utils.presets import MultiBackendRendererCfg
-
-from ... import dexsuite_env_cfg as dexsuite
-from ... import mdp
-
-FINGERTIP_LIST = ["panda_rightfinger", "panda_leftfinger"]
-
-
-BASE_CAMERA_CFG = CameraCfg(
-    prim_path="/World/envs/env_.*/Camera",
-    offset=CameraCfg.OffsetCfg(
-        pos=(0.57, -0.8, 0.5),
-        rot=(0.6124, 0.3536, 0.3536, 0.6124),
-        convention="opengl",
-    ),
-    data_types=MISSING,
-    spawn=sim_utils.PinholeCameraCfg(clipping_range=(0.01, 2.5)),
-    width=MISSING,
-    height=MISSING,
-    renderer_cfg=MultiBackendRendererCfg(),
-)
-
-WRIST_CAMERA_CFG = CameraCfg(
-    prim_path="/World/envs/env_.*/Robot/panda_hand/Camera",
-    offset=CameraCfg.OffsetCfg(
-        pos=(0.038, -0.38, -0.18),
-        rot=(0.641, 0.641, -0.299, 0.299),
-        convention="opengl",
-    ),
-    data_types=MISSING,
-    spawn=sim_utils.PinholeCameraCfg(clipping_range=(0.01, 2.5)),
-    width=MISSING,
-    height=MISSING,
-    renderer_cfg=MultiBackendRendererCfg(),
-)
-
-RAY_PATTERN = patterns.PinholeCameraPatternCfg(focal_length=24.0, horizontal_aperture=20.955)
-
-RAYCASTER_CAMERA_MESH_PRIM_PATHS = [
-    MultiMeshRayCasterCameraCfg.RaycastTargetCfg(
-        prim_expr="/World/envs/env_.*/table",
-        track_mesh_transforms=False,
-    ),
-    MultiMeshRayCasterCameraCfg.RaycastTargetCfg(
-        prim_expr="/World/GroundPlane",
-        track_mesh_transforms=False,
-    ),
-    MultiMeshRayCasterCameraCfg.RaycastTargetCfg(
-        prim_expr="/World/envs/env_.*/Object",
-        track_mesh_transforms=True,
-    ),
-    MultiMeshRayCasterCameraCfg.RaycastTargetCfg(
-        prim_expr="/World/envs/env_.*/Robot/.*/visuals",
-        track_mesh_transforms=True,
-    ),
-]
-
-BASE_RAYCASTER_CAMERA_CFG = MultiMeshRayCasterCameraCfg(
-    prim_path="/World/envs/env_.*/Camera",
-    offset=MultiMeshRayCasterCameraCfg.OffsetCfg(
-        pos=(0.57, -0.8, 0.5),
-        rot=(0.6124, 0.3536, 0.3536, 0.6124),
-        convention="opengl",
-    ),
-    mesh_prim_paths=RAYCASTER_CAMERA_MESH_PRIM_PATHS,
-    max_distance=2.5,
-    data_types=["distance_to_image_plane"],
-    pattern_cfg=MISSING,
-)
-
-WRIST_RAYCASTER_CAMERA_CFG = MultiMeshRayCasterCameraCfg(
-    prim_path="/World/envs/env_.*/Robot/panda_hand/Camera",
-    offset=MultiMeshRayCasterCameraCfg.OffsetCfg(
-        pos=(0.038, -0.38, -0.18),
-        rot=(0.641, 0.641, -0.299, 0.299),
-        convention="opengl",
-    ),
-    mesh_prim_paths=RAYCASTER_CAMERA_MESH_PRIM_PATHS,
-    max_distance=2.5,
-    data_types=["distance_to_image_plane"],
-    pattern_cfg=MISSING,
-    debug_vis=False,
-)
-
-
-@configclass
-class BaseTiledCameraCfg(PresetCfg):
-    """Tiled camera configurations"""
-
-    rgb64 = BASE_CAMERA_CFG.replace(data_types=["rgb"], width=64, height=64)
-    rgb128 = BASE_CAMERA_CFG.replace(data_types=["rgb"], width=128, height=128)
-    rgb256 = BASE_CAMERA_CFG.replace(data_types=["rgb"], width=256, height=256)
-    depth64 = BASE_CAMERA_CFG.replace(data_types=["depth"], width=64, height=64)
-    depth128 = BASE_CAMERA_CFG.replace(data_types=["depth"], width=128, height=128)
-    depth256 = BASE_CAMERA_CFG.replace(data_types=["depth"], width=256, height=256)
-    albedo64 = BASE_CAMERA_CFG.replace(data_types=["albedo"], width=64, height=64)
-    albedo128 = BASE_CAMERA_CFG.replace(data_types=["albedo"], width=128, height=128)
-    albedo256 = BASE_CAMERA_CFG.replace(data_types=["albedo"], width=256, height=256)
-    simple_shading_constant_diffuse64 = BASE_CAMERA_CFG.replace(
-        data_types=["simple_shading_constant_diffuse"], width=64, height=64
-    )
-    simple_shading_constant_diffuse128 = BASE_CAMERA_CFG.replace(
-        data_types=["simple_shading_constant_diffuse"], width=128, height=128
-    )
-    simple_shading_constant_diffuse256 = BASE_CAMERA_CFG.replace(
-        data_types=["simple_shading_constant_diffuse"], width=256, height=256
-    )
-    simple_shading_diffuse_mdl64 = BASE_CAMERA_CFG.replace(
-        data_types=["simple_shading_diffuse_mdl"], width=64, height=64
-    )
-    simple_shading_diffuse_mdl128 = BASE_CAMERA_CFG.replace(
-        data_types=["simple_shading_diffuse_mdl"], width=128, height=128
-    )
-    simple_shading_diffuse_mdl256 = BASE_CAMERA_CFG.replace(
-        data_types=["simple_shading_diffuse_mdl"], width=256, height=256
-    )
-    simple_shading_full_mdl64 = BASE_CAMERA_CFG.replace(data_types=["simple_shading_full_mdl"], width=64, height=64)
-    simple_shading_full_mdl128 = BASE_CAMERA_CFG.replace(data_types=["simple_shading_full_mdl"], width=128, height=128)
-    simple_shading_full_mdl256 = BASE_CAMERA_CFG.replace(data_types=["simple_shading_full_mdl"], width=256, height=256)
-    semantic_segmentation64 = BASE_CAMERA_CFG.replace(data_types=["semantic_segmentation"], width=64, height=64)
-    semantic_segmentation128 = BASE_CAMERA_CFG.replace(data_types=["semantic_segmentation"], width=128, height=128)
-    semantic_segmentation256 = BASE_CAMERA_CFG.replace(data_types=["semantic_segmentation"], width=256, height=256)
-    # raycaster camera presets
-    raycaster_depth64 = BASE_RAYCASTER_CAMERA_CFG.replace(pattern_cfg=RAY_PATTERN.replace(width=64, height=64))
-    raycaster_depth128 = BASE_RAYCASTER_CAMERA_CFG.replace(pattern_cfg=RAY_PATTERN.replace(width=128, height=128))
-    raycaster_depth256 = BASE_RAYCASTER_CAMERA_CFG.replace(pattern_cfg=RAY_PATTERN.replace(width=256, height=256))
-    default = rgb64
-
-
-@configclass
-class WristTiledCameraCfg(PresetCfg):
-    """Tiled camera configurations"""
-
-    rgb64 = WRIST_CAMERA_CFG.replace(data_types=["rgb"], width=64, height=64)
-    rgb128 = WRIST_CAMERA_CFG.replace(data_types=["rgb"], width=128, height=128)
-    rgb256 = WRIST_CAMERA_CFG.replace(data_types=["rgb"], width=256, height=256)
-    depth64 = WRIST_CAMERA_CFG.replace(data_types=["depth"], width=64, height=64)
-    depth128 = WRIST_CAMERA_CFG.replace(data_types=["depth"], width=128, height=128)
-    depth256 = WRIST_CAMERA_CFG.replace(data_types=["depth"], width=256, height=256)
-    albedo64 = WRIST_CAMERA_CFG.replace(data_types=["albedo"], width=64, height=64)
-    albedo128 = WRIST_CAMERA_CFG.replace(data_types=["albedo"], width=128, height=128)
-    albedo256 = WRIST_CAMERA_CFG.replace(data_types=["albedo"], width=256, height=256)
-    simple_shading_constant_diffuse64 = WRIST_CAMERA_CFG.replace(
-        data_types=["simple_shading_constant_diffuse"], width=64, height=64
-    )
-    simple_shading_constant_diffuse128 = WRIST_CAMERA_CFG.replace(
-        data_types=["simple_shading_constant_diffuse"], width=128, height=128
-    )
-    simple_shading_constant_diffuse256 = WRIST_CAMERA_CFG.replace(
-        data_types=["simple_shading_constant_diffuse"], width=256, height=256
-    )
-    simple_shading_diffuse_mdl64 = WRIST_CAMERA_CFG.replace(
-        data_types=["simple_shading_diffuse_mdl"], width=64, height=64
-    )
-    simple_shading_diffuse_mdl128 = WRIST_CAMERA_CFG.replace(
-        data_types=["simple_shading_diffuse_mdl"], width=128, height=128
-    )
-    simple_shading_diffuse_mdl256 = WRIST_CAMERA_CFG.replace(
-        data_types=["simple_shading_diffuse_mdl"], width=256, height=256
-    )
-    simple_shading_full_mdl64 = WRIST_CAMERA_CFG.replace(data_types=["simple_shading_full_mdl"], width=64, height=64)
-    simple_shading_full_mdl128 = WRIST_CAMERA_CFG.replace(data_types=["simple_shading_full_mdl"], width=128, height=128)
-    simple_shading_full_mdl256 = WRIST_CAMERA_CFG.replace(data_types=["simple_shading_full_mdl"], width=256, height=256)
-    semantic_segmentation64 = WRIST_CAMERA_CFG.replace(data_types=["semantic_segmentation"], width=64, height=64)
-    semantic_segmentation128 = WRIST_CAMERA_CFG.replace(data_types=["semantic_segmentation"], width=128, height=128)
-    semantic_segmentation256 = WRIST_CAMERA_CFG.replace(data_types=["semantic_segmentation"], width=256, height=256)
-    # raycaster camera presets
-    raycaster_depth64 = WRIST_RAYCASTER_CAMERA_CFG.replace(pattern_cfg=RAY_PATTERN.replace(width=64, height=64))
-    raycaster_depth128 = WRIST_RAYCASTER_CAMERA_CFG.replace(pattern_cfg=RAY_PATTERN.replace(width=128, height=128))
-    raycaster_depth256 = WRIST_RAYCASTER_CAMERA_CFG.replace(pattern_cfg=RAY_PATTERN.replace(width=256, height=256))
-    default = rgb64
-
-
-############################
-
-
-@configclass
-class StateObservationCfg(dexsuite.ObservationsCfg):
-    """Franka participant scene for Dexsuite Lifting/Reorientation"""
-
-    def __post_init__(self: dexsuite.ObservationsCfg):
-        super().__post_init__()
-        self.proprio.contact = ObsTerm(
-            func=mdp.fingers_contact_force_b,
-            params={"contact_sensor_names": [f"{link}_object_s" for link in FINGERTIP_LIST]},
-            clip=(-20.0, 20.0),  # contact force in finger tips is under 20N normally
-        )
-        self.proprio.hand_tips_state_b.params["body_asset_cfg"].body_names = FINGERTIP_LIST
-
-
-@configclass
-class SingleCameraObservationsCfg(StateObservationCfg):
-    """Observation specifications for the MDP."""
-
-    @configclass
-    class BaseImageObsCfg(ObsGroup):
-        """Camera observations for policy group."""
-
-        object_observation_b = ObsTerm(
-            func=mdp.vision_camera,
-            noise=Unoise(n_min=-0.0, n_max=0.0),
-            clip=(-1.0, 1.0),
-            params={"sensor_cfg": SceneEntityCfg("base_camera")},
-        )
-
-    base_image: BaseImageObsCfg = BaseImageObsCfg()
-
-    def __post_init__(self):
-        super().__post_init__()
-        for group in self.__dataclass_fields__.values():
-            obs_group = getattr(self, group.name)
-            obs_group.history_length = None
-
-
-@configclass
-class DuoCameraObservationsCfg(SingleCameraObservationsCfg):
-    """Observation specifications for the MDP."""
-
-    @configclass
-    class WristImageObsCfg(ObsGroup):
-        wrist_observation = ObsTerm(
-            func=mdp.vision_camera,
-            noise=Unoise(n_min=-0.0, n_max=0.0),
-            clip=(-1.0, 1.0),
-            params={"sensor_cfg": SceneEntityCfg("wrist_camera")},
-        )
-
-    wrist_image: WristImageObsCfg = WristImageObsCfg()
